refactor: move event join function
유저가 모임에 참여
userEventController -> EventController 이동
